Refactor application apply
