+++ application/forms/__init__.py
@@ -1,0 +1,4 @@
+from .hacker import HackerForm
+from .volunteer import VolunteerForm
+from .mentor import MentorForm
+from .sponsor import SponsorForm
