+++ application/templates/application_home/apply.html
@@ -7,13 +7,13 @@
     <ul>
         {% for apply_type in apply_types %}
             <li>
-                {{ apply_type.get_description }}<br>
+                <strong>{{ apply_type.name }}: </strong>{{ apply_type.get_description }}<br>
                 <strong>
                     {% if apply_type.active %}{% translate 'Closes in' %}{% else %}{% translate 'Opens in' %}{% endif %}{{ apply_type.time_left }}
                 </strong>
                 <div class="row justify-content-around mt-3 mb-3">
                     <div class="col-6 d-grid d-md-block">
-                        <a href="{% url 'apply' %}?type={{ apply_type.name }}" class="btn {% if user_applications %}btn-secondary{% else %}btn-primary{% endif %} col-12">{% trans 'Apply as' %} {{ apply_type.name }}</a>
+                        <a href="{% url 'apply' %}?type={{ apply_type.name }}&token={{ apply_type.get_token }}" class="btn {% if user_applications %}btn-secondary{% else %}btn-primary{% endif %} col-12">{% trans 'Apply as' %} {{ apply_type.name }}</a>
                     </div>
                 </div>
             </li>
