+++ review/templates/application_list.html
@@ -21,7 +21,32 @@
                 </div>
             {% endif %}
         </div>
-        <h1 class="mt-3">{% translate 'Application list' %}</h1>
+        <div class="row mt-3">
+            <div class="col-9">
+                <h1>{% translate 'Application list' %}</h1>
+            </div>
+            <div class="col-3 text-end">
+                <button type="button" class="btn btn-secondary align-middle" data-bs-toggle="modal" data-bs-target="#urlFormLink"><i class="bi bi-info-circle-fill"></i></button>
+            </div>
+            <div class="modal fade" id="urlFormLink" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
+                <div class="modal-dialog">
+                    <div class="modal-content bg-{{ theme }} border-1">
+                        <div class="modal-header">
+                            <h5 class="modal-title" id="exampleModalLabel">{% translate 'Application form link' %}</h5>
+                            <button type="button" class="btn-close text-white" data-bs-dismiss="modal" aria-label="Close"></button>
+                        </div>
+                        <div class="modal-body">
+                            <input id="url-copy" class="form-control" value="{{ apply_url }}?type={{ application_type.name }}&token={{ application_type.get_token }}"/>
+                            <div class="row justify-content-center mt-2">
+                                <div class="col-12 col-lg-6 d-grid d-md-block">
+                                    <button type="button" class="btn btn-primary col-12" onclick="copyUrl()">{% translate 'Copy link' %}</button>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
     {% else %}
         <div class="row justify-content-between gy-2">
             <div class="col-12 col-lg-2 d-grid d-md-block">
@@ -38,6 +63,20 @@
     <form method="get" id="filter-form" class="mt-2">
         {% include 'components/bootstrap5_form.html' with form=filter.form is_filter=True %}
     </form>
+
+    <script>
+        function copyUrl() {
+              // Get the text field
+              let copyText = document.getElementById("url-copy");
+
+              // Select the text field
+              copyText.select();
+              copyText.setSelectionRange(0, 99999); // For mobile devices
+
+               // Copy the text inside the text field
+              navigator.clipboard.writeText(copyText.value);
+            }
+    </script>
 
     {% if not invite %}
         {% include 'components/table.html' %}
