+++ user/templates/password_reset.html
@@ -9,23 +9,46 @@
             <div class="bg-{{ theme }} text-{% if theme == 'dark' %}white{% else %}black{% endif %}">
                 <div class="content p-4">
                     {% if success %}
-                        <h1>{% translate 'Password reset complete' %}</h1>
+                        <h1>
+                            {% if not new %}
+                                {% translate 'Password reset complete' %}
+                            {% else %}
+                                {% translate 'Password complete' %}
+                            {% endif %}
+                        </h1>
 
                         <p>
-                            {% translate 'Your password has been set.' %}
+                            {% if not new %}
+                                {% translate 'Your password has been set.' %}
+                            {% else %}
+                                {% translate 'Now you have a user to enter to the app. Verify your email to finish.' %}
+                            {% endif %}
+
                         </p>
                         <div class="d-grid col-6 mx-auto">
                             <a class="btn btn-primary" href="{% url 'login' %}">{% trans "Log in" %}</a>
                         </div>
                     {% else %}
-                        <h1>{% translate 'Forgot password' %}</h1>
+                        <h1>
+                            {% if not new %}
+                                {% translate 'Forgot password' %}
+                            {% else %}
+                                {% translate 'Set password' %}
+                            {% endif %}
+                        </h1>
 
                         <form method="post">
                             {% csrf_token %}
                             {% include 'components/bootstrap5_form.html' %}
 
                             <div class="d-grid col-6 mx-auto">
-                                <button class="btn btn-primary" type="submit">{% trans "Reset password" %}</button>
+                                <button class="btn btn-primary" type="submit">
+                                    {% if not new %}
+                                        {% translate 'Forgot password' %}
+                                    {% else %}
+                                        {% translate 'Set password' %}
+                                    {% endif %}
+                                </button>
                             </div>
                         </form>
                     {% endif %}
